// frontend/src/api_client/models/PickupStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum PickupStatus {
    DRAFT = 'DRAFT',
    OPEN = 'OPEN',
    ASSIGNED = 'ASSIGNED',
    IN_TRANSIT = 'IN_TRANSIT',
    COMPLETED = 'COMPLETED',
    CANCELLED = 'CANCELLED',
    RTO_INITIATED = 'RTO_INITIATED',
}